+++ src/main/java/club/inq/team1/dto/request/UpdateUserPasswordDTO.java
@@ -1,4 +1,4 @@
-package club.inq.team1.dto;
+package club.inq.team1.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
