+++ src/main/java/club/inq/team1/dto/request/UserJoinDTO.java
@@ -1,4 +1,4 @@
-package club.inq.team1.dto;
+package club.inq.team1.dto.request;
 
 import club.inq.team1.config.Gender;
 import io.swagger.v3.oas.annotations.media.Schema;
